+++ utils/libraries.js
@@ -53,37 +53,7 @@ const kvLibraryVersion = async (library, version) => {
     return data;
 };
 
-// Validate the data we get from KV for a library
-const kvLibraryValidate = (library, data) => {
-    // Assets might not exist if there are none, but we should make it an empty array by default
-    data.assets = data.assets || [];
 
-    // Non-breaking issues
-    if (library !== data.name) {
-        console.info('Name mismatch', library, data.name);
-        data.name = library;
-    }
-
-    // Breaking issues
-    if (!data.version) {
-        console.error('Version missing', data.name, data);
-        if (process.env.SENTRY_DSN) {
-            Sentry.withScope(scope => {
-                scope.setTag('data', JSON.stringify(data));
-                Sentry.captureException(new Error('Version missing in package data'));
-            });
-        }
-        throw new Error('Version missing in package data');
-    }
-
-    return data;
-};
-
-// Get the metadata for a library on KV
-const kvLibrary = async library => {
-    const data = await jsonCachedFetch(`${kvBase}/packages/${encodeURIComponent(library)}`);
-    return kvLibraryValidate(library, data);
-};
 
 // Get the full metadata for a library incl. versions on KV
 const kvFullLibrary = async library => {
